# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
      "Answer" : {
        "properties" : {
          "answerId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "authorCoarseRegionCode" : {
            "type" : "string"
          },
          "authorNickname" : {
            "type" : "string"
          },
          "bearingFromSenderDegrees" : {
            "type" : "number"
          },
          "bodyText" : {
            "type" : "string"
          },
          "distanceBand" : {
            "type" : "string"
          },
          "distanceM" : {
            "format" : "int64",
            "type" : "integer"
          },
          "editedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "mediaIds" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reactedByMe" : {
            "type" : "boolean"
          },
          "reactionCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "AnswerListingResponse" : {
        "properties" : {
          "answers" : {
            "items" : {
              "$ref" : "#/components/schemas/Answer"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "$ref" : "#/components/schemas/Cursor"
          }
        },
        "type" : "object"
      },
...
      "AnswersReadResponse" : {
        "properties" : {
          "answersReadAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiErrorDetail" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "field" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiErrorResponse" : {
        "properties" : {
          "errorDetail" : {
            "$ref" : "#/components/schemas/ApiErrorDetail"
          },
          "message" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [
              "success",
              "error"
            ],
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseAnswerListingResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/AnswerListingResponse"
          },
          "status" : {
            "enum" : [
              "success",
              "error"
            ],
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseAnswersReadResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/AnswersReadResponse"
          },
          "status" : {
            "enum" : [
              "success",
              "error"
            ],
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Cursor" : {
        "properties" : {
          "postId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "submittedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "appAccessToken" : {
        "bearerFormat" : "JWT",
        "description" : "앱 액세스 토큰. 기기 등록 또는 토큰 재발급으로 받는다.",
        "scheme" : "bearer",
        "type" : "http"
      },
      "operatorSession" : {
        "description" : "백오피스 세션 쿠키. POST /admin/login 성공 시 발급된다.",
        "in" : "cookie",
        "name" : "SESSION",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "description" : "Qello 백엔드 API. 모든 응답은 docs/api-response.md의 공통 계약을 따른다.\n성공은 status/data/timestamp, 실패는 status/message/errorDetail/timestamp를 갖는다.\n오류 코드 목록은 docs/error-codes.md에 있다.",
    "title" : "Qello API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/direction/inbox/{postRecipientId}/answers/read" : {
      "put" : {
        "description" : "인증 사용자가 받은 질문의 답변 열람 시각을 서버 시각으로 기록합니다. 반복 호출은 이미 기록된 시각보다 이전으로 되돌리지 않습니다.",
        "operationId" : "markRecipientAnswersRead",
        "parameters" : [
          {
            "description" : "수신함 항목 식별자",
            "example" : 101,
            "in" : "path",
            "name" : "postRecipientId",
            "required" : true,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAnswersReadResponse"
                }
              }
            },
            "description" : "갱신된 읽음 시각을 반환합니다."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "요청 값이 올바르지 않습니다. 오류 코드는 docs/error-codes.md를 따른다."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "앱 액세스 토큰이 유효하지 않습니다."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "현재 계정은 이 기능을 사용할 수 없습니다. (FED-APP-002)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "수신 항목을 찾을 수 없습니다. (DIR-DOM-008)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "서버 내부 오류가 발생했습니다."
          }
        },
        "security" : [
          {
            "appAccessToken" : [ ]
          }
        ],
        "summary" : "수신자의 답변 읽음 처리",
        "tags" : [
          "답변 읽음 처리"
        ]
      }
    },
...
    "/api/v1/direction/posts/{postId}/answers" : {
      "get" : {
        "description" : "그 질문글의 공개된 답변을 최신순으로 조회합니다. 질문글 작성자와 그 질문글의 수신 자격자만 내용을 받습니다. 자격이 없는 뷰어는 403이 아니라 빈 목록을 받습니다 — 질문글 존재 여부를 흘리지 않기 위함입니다.",
        "operationId" : "answers",
        "parameters" : [
          {
            "description" : "질문글 식별자",
            "example" : 101,
            "in" : "path",
            "name" : "postId",
            "required" : true,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          },
          {
            "in" : "query",
            "name" : "cursorPublishedAt",
            "required" : false,
            "schema" : {
              "format" : "date-time",
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "cursorAnswerId",
            "required" : false,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          },
          {
            "in" : "query",
            "name" : "limit",
            "required" : false,
            "schema" : {
              "default" : 20,
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAnswerListingResponse"
                }
              }
            },
            "description" : "답변 목록을 반환합니다. 자격이 없으면 빈 목록입니다."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "limit 또는 cursor 파라미터가 올바르지 않습니다. (FED-VAL-001, FED-VAL-002)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "앱 액세스 토큰이 유효하지 않습니다."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "현재 계정은 이 기능을 사용할 수 없습니다. (FED-APP-002)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "인증 사용자 계정을 찾을 수 없습니다. (FED-APP-001)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "서버 내부 오류가 발생했습니다."
          }
        },
        "security" : [
          {
            "appAccessToken" : [ ]
          }
        ],
        "summary" : "답변 목록 조회",
        "tags" : [
          "내가 보낸 질문"
        ]
      }
    },
    "/api/v1/direction/posts/{postId}/answers/read" : {
      "put" : {
        "description" : "인증 사용자가 보낸 질문글의 답변 열람 시각을 서버 시각으로 기록합니다. 반복 호출은 이미 기록된 시각보다 이전으로 되돌리지 않습니다.",
        "operationId" : "markSenderAnswersRead",
        "parameters" : [
          {
            "description" : "질문글 식별자",
            "example" : 101,
            "in" : "path",
            "name" : "postId",
            "required" : true,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAnswersReadResponse"
                }
              }
            },
            "description" : "갱신된 읽음 시각을 반환합니다."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "요청 값이 올바르지 않습니다. 오류 코드는 docs/error-codes.md를 따른다."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "앱 액세스 토큰이 유효하지 않습니다."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "현재 계정은 이 기능을 사용할 수 없습니다. (FED-APP-002)"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "질문글을 찾을 수 없습니다. (DIR-DOM-009)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "서버 내부 오류가 발생했습니다."
          }
        },
        "security" : [
          {
            "appAccessToken" : [ ]
          }
        ],
        "summary" : "질문자의 답변 읽음 처리",
        "tags" : [
          "답변 읽음 처리"
        ]
      }
    },
...
  },
  "servers" : [
    {
      "description" : "Generated server url",
      "url" : "http://localhost"
    }
  ],
...
}